Refactor saveItem test
